+++ structurizr-core/test/unit/com/structurizr/view/SequenceNumberTests.java
@@ -14,32 +14,19 @@ public class SequenceNumberTests {
     }
 
     @Test
-    public void test_childSequence() {
-        SequenceNumber sequenceNumber = new SequenceNumber();
-        assertEquals("1", sequenceNumber.getNext());
-
-        sequenceNumber.startChildSequence();
-        assertEquals("1.1", sequenceNumber.getNext());
-        assertEquals("1.2", sequenceNumber.getNext());
-
-        sequenceNumber.endChildSequence();
-        assertEquals("2", sequenceNumber.getNext());
-    }
-
-    @Test
     public void test_parallelSequences() {
         SequenceNumber sequenceNumber = new SequenceNumber();
         assertEquals("1", sequenceNumber.getNext());
 
         sequenceNumber.startParallelSequence();
         assertEquals("2", sequenceNumber.getNext());
-        sequenceNumber.endParallelSequence();
+        sequenceNumber.endParallelSequence(false);
 
         sequenceNumber.startParallelSequence();
         assertEquals("2", sequenceNumber.getNext());
-        sequenceNumber.endParallelSequence();
+        sequenceNumber.endParallelSequence(true);
 
-        assertEquals("2", sequenceNumber.getNext());
+        assertEquals("3", sequenceNumber.getNext());
     }
 
 }